apps/api/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# AI/ML and RAG
openai==1.10.0
langchain==0.1.5
langchain-openai==0.0.6
qdrant-client==1.7.3
sentence-transformers==2.2.2
pypdf2==3.0.1
youtube-transcript-api==0.6.1
pytube==15.0.0

# AWS
boto3==1.34.34
botocore==1.34.34

# Validation and serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client
httpx==0.26.0
requests==2.31.0

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
black==24.1.1
isort==5.13.2
flake8==7.0.0

# Utilities
python-dotenv==1.0.1
redis==5.0.1